finish correct wiring of graph enablement and configuration for a given resource type
